raspberry_pi_3b/02_room_buddy/requirements.txt
Adafruit-Blinka==8.50.0
adafruit-circuitpython-busdevice==5.2.10
adafruit-circuitpython-connectionmanager==3.1.2
adafruit-circuitpython-dht==4.0.6
adafruit-circuitpython-requests==4.1.8
adafruit-circuitpython-typing==1.11.2
Adafruit-PlatformDetect==3.75.0
Adafruit-PureIO==1.1.11
aiohappyeyeballs==2.4.4
aiohttp==3.11.10
aiosignal==1.3.2
arrow==1.3.0
attrs==24.2.0
binho-host-adapter==0.1.6
cairocffi==1.7.1
CairoSVG==2.7.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
colorzero==2.0
cssselect2==0.7.0
defusedxml==0.7.1
frozenlist==1.5.0
gpiozero==2.0.1
ics==0.7.2
idna==3.10
lgpio==0.2.2.0
multidict==6.1.0
pillow==11.0.0
propcache==0.2.1
pycparser==2.22
pyftdi==0.56.0
pyserial==3.5
python-dateutil==2.9.0.post0
pyusb==1.2.1
qrcode==8.0
RPi.GPIO==0.7.1
rpi_ws281x==5.0.0
six==1.17.0
spidev==3.6
systemd-python==235
sysv_ipc==1.1.0
TatSu==5.12.2
tinycss2==1.4.0
types-python-dateutil==2.9.0.20241206
typing_extensions==4.12.2
urllib3==2.2.3
webencodings==0.5.1
yarl==1.18.3
